requirements.txt
flask>=3.0.0
werkzeug>=3.0.0
gunicorn>=21.0.0
opentelemetry-api>=1.27.0
opentelemetry-sdk>=1.27.0
opentelemetry-exporter-otlp>=1.27.0
opentelemetry-instrumentation-flask>=0.48b0
opentelemetry-instrumentation-sqlite3>=0.48b0
opentelemetry-semantic-conventions>=0.48b0
